+++ src/Config/Option/Database/DatabaseUserConfigOption.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Installer\Config\Option\Database;
+
+use Shlinkio\Shlink\Config\Collection\PathCollection;
+use Shlinkio\Shlink\Installer\Util\AskUtilsTrait;
+use Symfony\Component\Console\Style\StyleInterface;
+
+class DatabaseUserConfigOption extends AbstractNonSqliteDependentConfigOption
+{
+    use AskUtilsTrait;
+
+    public function getConfigPath(): array
+    {
+        return ['entity_manager', 'connection', 'user'];
+    }
+
+    public function ask(StyleInterface $io, PathCollection $currentOptions): string
+    {
+        return $this->askRequired($io, 'username', 'Database username');
+    }
+}