+++ src/Config/Option/Visit/VisitsWebhooksConfigOption.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Installer\Config\Option\Visit;
+
+use Shlinkio\Shlink\Config\Collection\PathCollection;
+use Shlinkio\Shlink\Installer\Config\Option\AbstractSwooleDependentConfigOption;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidatorsTrait;
+use Symfony\Component\Console\Style\StyleInterface;
+
+class VisitsWebhooksConfigOption extends AbstractSwooleDependentConfigOption
+{
+    use ConfigOptionsValidatorsTrait;
+
+    public function ask(StyleInterface $io, PathCollection $currentOptions): array
+    {
+        return $io->ask(
+            'Provide a comma-separated list of webhook URLs which will receive POST notifications when short URLs '
+            . 'receive visits (Ignore this if you are not serving shlink with swoole)',
+            null,
+            [$this, 'splitAndValidateMultipleUrls'],
+        );
+    }
+
+    public function getConfigPath(): array
+    {
+        return ['url_shortener', 'visits_webhooks'];
+    }
+}