+++ src/Config/Option/Worker/AbstractWorkerNumConfigOption.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Installer\Config\Option\Worker;
+
+use Shlinkio\Shlink\Config\Collection\PathCollection;
+use Shlinkio\Shlink\Installer\Config\Option\AbstractSwooleDependentConfigOption;
+use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidatorsTrait;
+use Symfony\Component\Console\Style\StyleInterface;
+
+use function sprintf;
+
+abstract class AbstractWorkerNumConfigOption extends AbstractSwooleDependentConfigOption
+{
+    use ConfigOptionsValidatorsTrait;
+
+    public function ask(StyleInterface $io, PathCollection $currentOptions): int
+    {
+        $question = sprintf('%s (Ignore this if you are not serving shlink with swoole)', $this->getQuestionToAsk());
+        return $io->ask($question, '16', [$this, 'validatePositiveNumber']);
+    }
+
+    abstract protected function getQuestionToAsk(): string;
+}